move null library tests to a separare file
